requirements.txt
streamlit
google-generativeai
feedparser